Ripped the allowed_hosts module out into separate module files
