+++ service-providers/network-requester/src/allowed_hosts/mod.rs
@@ -1,0 +1,6 @@
+mod filter;
+mod host;
+mod storage;
+
+pub(crate) use filter::OutboundRequestFilter;
+pub(crate) use storage::HostsStore;
