+++ service-providers/network-requester/src/allowed_hosts/storage.rs
@@ -1,0 +1,122 @@
+use std::{
+    collections::HashSet,
+    fs::{self, File, OpenOptions},
+    io::{self, BufRead, BufReader},
+    net::IpAddr,
+    path::{Path, PathBuf},
+};
+
+use super::host::Host;
+use ipnetwork::IpNetwork;
+
+/// A simple file-based store for information about allowed / unknown hosts.
+/// Currently it completely ignores any port information.
+// TODO: in the future allow filtering by port, so for example 1.1.1.1:80 would be a valid filter,
+// which would allow connections to the port :80 while any requests to say 1.1.1.1:1234 would be denied.
+#[derive(Debug)]
+pub(crate) struct HostsStore {
+    pub(super) storefile: PathBuf,
+
+    pub(super) domains: HashSet<String>,
+    pub(super) ip_nets: Vec<IpNetwork>,
+}
+
+impl HostsStore {
+    /// Constructs a new HostsStore
+    pub(crate) fn new(base_dir: PathBuf, filename: PathBuf) -> HostsStore {
+        let storefile = HostsStore::setup_storefile(base_dir, filename);
+        let hosts = HostsStore::load_from_storefile(&storefile)
+            .unwrap_or_else(|_| panic!("Could not load hosts from storefile at {:?}", storefile));
+
+        let (domains, ip_nets): (Vec<_>, Vec<_>) =
+            hosts.into_iter().partition(|host| host.is_domain());
+
+        HostsStore {
+            storefile,
+            domains: domains.into_iter().map(Host::extract_domain).collect(),
+            ip_nets: ip_nets.into_iter().map(Host::extract_ipnetwork).collect(),
+        }
+    }
+
+    pub(super) fn append(path: &Path, text: &str) {
+        use std::io::Write;
+        let mut file = OpenOptions::new()
+            .write(true)
+            .append(true)
+            .open(path)
+            .unwrap();
+
+        if let Err(e) = writeln!(file, "{}", text) {
+            eprintln!("Couldn't write to file: {}", e);
+        }
+    }
+
+    fn append_to_file(&self, host: &str) {
+        HostsStore::append(&self.storefile, host);
+    }
+
+    pub(crate) fn contains_domain(&self, host: &str) -> bool {
+        self.domains.contains(&host.to_string())
+    }
+
+    pub(super) fn contains_ip_address(&self, address: IpAddr) -> bool {
+        // I'm not sure it's possible to achieve the same functionality without iterating through
+        // the whole thing. Maybe by some clever usage of tries? But I doubt we're going to have
+        // so many filtering rules that it's going to matter at this point.
+        for ip_net in &self.ip_nets {
+            if ip_net.contains(address) {
+                return true;
+            }
+        }
+
+        false
+    }
+
+    /// Returns the default base directory for the storefile.
+    ///
+    /// This is split out so we can easily inject our own base_dir for unit tests.
+    pub fn default_base_dir() -> PathBuf {
+        dirs::home_dir()
+            .expect("no home directory known for this OS")
+            .join(".nym")
+    }
+
+    pub(super) fn maybe_add_ip(&mut self, ip: IpAddr) {
+        if !self.contains_ip_address(ip) {
+            self.ip_nets.push(ip.into());
+            self.append_to_file(&ip.to_string());
+        }
+    }
+
+    pub(super) fn maybe_add_domain(&mut self, domain: &str) {
+        if !self.contains_domain(domain) {
+            self.domains.insert(domain.to_string());
+            self.append_to_file(domain);
+        }
+    }
+
+    fn setup_storefile(base_dir: PathBuf, filename: PathBuf) -> PathBuf {
+        let dirpath = base_dir.join("service-providers").join("network-requester");
+        fs::create_dir_all(&dirpath)
+            .unwrap_or_else(|_| panic!("could not create storage directory at {:?}", dirpath));
+        let storefile = dirpath.join(filename);
+        let exists = std::path::Path::new(&storefile).exists();
+        if !exists {
+            File::create(&storefile).unwrap();
+        }
+        storefile
+    }
+
+    /// Loads the storefile contents into memory.
+    pub(super) fn load_from_storefile<P>(filename: P) -> io::Result<Vec<Host>>
+    where
+        P: AsRef<Path>,
+    {
+        let file = File::open(filename)?;
+        let reader = BufReader::new(&file);
+        Ok(reader
+            .lines()
+            .map(|line| Host::from(line.expect("failed to read input file line!")))
+            .collect())
+    }
+}
